AP_Math: split CRCs from checksums
All CRCs are checksums but not all checksums are CRCs.

Paves the way for deduplication and unification of CRC implementations,
including interface and standard of implementation.
