+++ libraries/AP_Math/crc.cpp
@@ -13,7 +13,7 @@
   along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 /*
-  collection of CRCs. 
+  collection of CRC calculation functions
  */
 
 #include <stdint.h>
@@ -482,29 +482,6 @@ uint16_t calc_crc_modbus(const uint8_t *buf, uint16_t len)
     return crc;
 }
 
-// fletcher 16 implementation
-uint16_t crc_fletcher16(const uint8_t *buffer, uint32_t len) {
-    uint16_t c0 = 0;
-    uint16_t c1 = 0;
-    for (uint32_t i = 0; i < len; i++) {
-        c0 = (c0 + buffer[i]) % 255;
-        c1 = (c1 + c0) % 255;
-    }
-
-    return (c1 << 8) | c0;
-}
-
-// FNV-1a implementation
-#define FNV_1_PRIME_64 1099511628211UL
-void hash_fnv_1a(uint32_t len, const uint8_t* buf, uint64_t* hash)
-{
-    uint32_t i;
-    for (i=0; i<len; i++) {
-        *hash ^= (uint64_t)buf[i];
-        *hash *= FNV_1_PRIME_64;
-    }
-}
-
 // calculate 24 bit crc. We take an approach that saves memory and flash at the cost of higher CPU load.
 uint32_t crc_crc24(const uint8_t *bytes, uint16_t len)
 {
@@ -523,19 +500,6 @@ uint32_t crc_crc24(const uint8_t *bytes, uint16_t len)
         crc = ((crc<<8)&0xFFFFFF) ^ crct;
     }
     return crc;
-}
-
-// simple 8 bit checksum used by FPort
-uint8_t crc_sum8_with_carry(const uint8_t *p, uint8_t len)
-{
-    uint16_t sum = 0;
-    for (uint8_t i=0; i<len; i++) {
-        sum += p[i];
-        sum += sum >> 8;
-        sum &= 0xFF;              
-    }
-    sum = 0xff - ((sum & 0xff) + (sum >> 8));
-    return sum;
 }
 
 // CRC-16 (IBM/ANSI)
@@ -615,48 +579,4 @@ uint64_t crc_crc64(const uint32_t *data, uint16_t num_words)
     crc ^= ~(0ULL);
 
     return crc;
-}
-
-// return the parity of byte - "1" if there is an odd number of bits
-// set, "0" if there is an even number of bits set note that
-// __builtin_parity causes hardfaults on Pixracer-periph - and is
-// slower on 1 byte than this:
-uint8_t parity(uint8_t byte)
-{
-    uint8_t p = 0;
-
-    p ^= byte & 0x1;
-    byte >>= 1;
-    p ^= byte & 0x1;
-    byte >>= 1;
-    p ^= byte & 0x1;
-    byte >>= 1;
-    p ^= byte & 0x1;
-    byte >>= 1;
-    p ^= byte & 0x1;
-    byte >>= 1;
-    p ^= byte & 0x1;
-    byte >>= 1;
-    p ^= byte & 0x1;
-    byte >>= 1;
-    p ^= byte & 0x1;
-
-    return p;
-}
-
-// sums the bytes in the supplied buffer, returns that sum mod 0xFFFF
-uint16_t crc_sum_of_bytes_16(const uint8_t *data, uint16_t count)
-{
-    uint16_t ret = 0;
-    for (uint32_t i=0; i<count; i++) {
-        ret += data[i];
-    }
-    return ret;
-}
-
-// sums the bytes in the supplied buffer, returns that sum mod 256
-// (i.e. shoved into a uint8_t)
-uint8_t crc_sum_of_bytes(const uint8_t *data, uint16_t count)
-{
-    return crc_sum_of_bytes_16(data, count) & 0xFF;
 }
